+++ config/deploy/staging.rb
@@ -1,5 +1,5 @@
 set :stage, :staging
-set :branch, "develop"
+set :branch, "more-translations"
 
 server "web-supported-staging.linode.unep-wcmc.org", user: 'wcmc', roles: %w{app web db}
